% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{yar-package}
\alias{yar}
\alias{yar-package}
\title{yar: Y-CRDT Bindings to R}
\description{
Provides R bindings to y-crdt, a Rust implementation of the Yjs CRDT (Conflict-free Replicated Data Type) framework.
}
\author{
\strong{Maintainer}: Antoine Prouvost \email{AntoinePrv@users.noreply.github.com}

}
\keyword{internal}
